Rename 'repo' package to 'repository' for clarity.
Updated package naming to improve readability and consistency. All references and imports were adjusted to reflect the new 'repository' package structure. No functional changes were made.
