+++ src/main/java/com/edu/pharmacy/repository/MedicineRepository.java
@@ -1,4 +1,4 @@
-package com.edu.pharmacy.repo;
+package com.edu.pharmacy.repository;
 
 import com.edu.pharmacy.entity.MedicineEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
