+++ src/main/java/com/edu/pharmacy/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.edu.pharmacy.repo;
+package com.edu.pharmacy.repository;
 
 import com.edu.pharmacy.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
